requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
apscheduler>=3.10.4
requests>=2.31.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
numpy>=1.26.0
